# Further Optimization and Speedup