--- a/NVC_COMPILE
+++ b/NVC_COMPILE
@@ -8,16 +8,16 @@ LINKFLAG="-L/opt/nvidia/hpc_sdk/Linux_x86_64/22.5/cuda/lib64 -L/usr/lib64/ -L/op
 
 NVCFLAG="-O3 -std=c++20 -target=gpu -Minline -fopenmp -cuda -stdpar=multicore"
 
-$CXX $NVCFLAG $LINKFLAG -c axpy.cu
+$CXX $NVCFLAG $LINKFLAG -c axpy.cu ; echo "Finished axpy.cu"
 
-$CXX $NVCFLAG $LINKFLAG -c main.cpp
+$CXX $NVCFLAG $LINKFLAG -c main.cpp ; echo "Finished main.cpp"
 
-$CXX $NVCFLAG $LINKFLAG -c read_data.cpp
+$CXX $NVCFLAG $LINKFLAG -c read_data.cpp ; echo "Finished read_data.cpp"
 
-$CXX $NVCFLAG $LINKFLAG -c data_struct.cpp
+$CXX $NVCFLAG $LINKFLAG -c data_struct.cpp ; echo "Finished data_struct.cpp"
 
-$CXX $NVCFLAG $LINKFLAG -c VDW_Coulomb.cu
+$CXX $NVCFLAG $LINKFLAG -c VDW_Coulomb.cu ; echo "Finished VDW_Coulomb.cu"
 
-$CXX $NVCFLAG $LINKFLAG main.o read_data.o axpy.o data_struct.o VDW_Coulomb.o -o nvc_main.x
+$CXX $NVCFLAG $LINKFLAG main.o read_data.o axpy.o data_struct.o VDW_Coulomb.o -o nvc_main.x ; echo "Finished Linking"
 
 rm *.o