1. Further Optimized Ewald Summation Initialization, speed is increased by 10-20% for simulations
   * Examples:
     * CO2-MFI: 9.7 → 8.3 seconds
     * Bae-Mixture: 108.8 → 97.2 seconds
     * NU-2000: 17.20 → 7.97 seconds
     * NPTMC: 223.73 → 180.81 seconds
     * NVT-Gibbs (100 + 100 cycles): 35.3 → 31.82 seconds
2. Small update for Compilation file: NVC_COMPILE
   * added echoes